Add support for nonce for inline scripts
If an application is using a Content Security Policy (CSP)
then inline scripts will fail the policy and cause
the JavaScript not to run.

To avoid this we can use a hash of the inline script, but since
the inline script will change depending on the CDN Host and
what version of GOV.UK Frontend is being used a nonce attribute
can be used instead.

See https://content-security-policy.com/examples/allow-inline-script/
for more info.